backend/requirements.txt
FastAPI>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
redis==4.3.4
celery==5.2.7
python-jose==3.3.0
bcrypt==3.2.0
passlib>=1.7.4
requests==2.26.0
pymongo==3.12.1
elasticsearch==7.16.0
prometheus-fastapi-instrumentator>=5.0.0
logstash-formatter>=0.5.17